requirements.txt
fastapi
uvicorn
streamlit
openai
python-dotenv
httpx
python-multipart
requests 